не хватет ошибок

+++ src/main/java/dao/api/IDaoCurrency.java
@@ -5,9 +5,9 @@ import core.Currency;
 import java.util.List;
 
 public interface IDaoCurrency {
-    int getID(String typeCurrency);
+    long getID(String typeCurrency);
     void saveCurrency(List<Currency> currencies);
     boolean exist(String typeCurrency);
     List<Currency> getCurrency();
-    void remove(int id);
+    void remove(long id);
 }
